requirements.txt
azure-identity>=1.15.0
azure-mgmt-costmanagement>=4.0.0
azure-mgmt-resource>=23.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0

